remove valence_service_integration & balancer mods; toml updates
