+++ Cargo.toml
@@ -49,7 +49,7 @@ serde            = { version = "1.0.207", default-features = false, features = [
 serde_json       = "1.0.125"
 sha2             = "0.10.8"
 thiserror        = "1.0.63"
-osmosis-std     = "0.26.0"
+osmosis-std      = "0.26.0"
 
 # our contracts
 valence-authorization            = { path = "contracts/authorization", features = ["library"] }
@@ -86,4 +86,4 @@ hex                  = "0.4.3"
 margined-neutron-std = "4.2.0"
 neutron-test-tube    = "4.2.0"
 tokio                = "1.40.0"
-osmosis-test-tube    = { version = "25.0.0" }
+osmosis-test-tube    = "25.0.0"
