+++ contracts/services/osmosis-gamm-lper/src/bin/schema.rs
@@ -1,9 +1,6 @@
 use cosmwasm_schema::write_api;
 
-use valence_osmosis_gamm_lper::{
-    msg::{ActionsMsgs, QueryMsg},
-    valence_service_integration::{OptionalServiceConfig, ServiceConfig},
-};
+use valence_osmosis_gamm_lper::msg::{ActionsMsgs, OptionalServiceConfig, QueryMsg, ServiceConfig};
 use valence_service_utils::msg::{ExecuteMsg, InstantiateMsg};
 
 fn main() {
