+++ contracts/services/osmosis-gamm-lper/src/lib.rs
@@ -1,6 +1,4 @@
-mod balancer;
 pub mod contract;
 pub mod msg;
 #[cfg(test)]
 mod testing;
-pub mod valence_service_integration;
